handle models which are plain dicts
A model of a generic type (i.e. an object) does not have a
specification. Consequently, it does not have the attributes of a
model object such as composed_schemas. Handle this case by checking for
it and short-circuiting the model_instances loop early in th
model_to_dict function. Use the same code as the dict-case in that loop
since they are fundamentally the same case.

TL;DR: we hanlde the case of a dict embedded within a dict, whereas
previously we did not. There is likely a more fundamental problem here
(what if the attribute was a list?) that requires greater rework. But,
we're running in to this now, so fix it now.